+++ apps/manager/app/league/_components/LeagueForm/index.tsx
@@ -1,0 +1,2 @@
+export * from './LeagueForm';
+export * from './types';
